Added dialogs for Watchlist importer
Added dialogs for Watchlist importer for various circumstances:

1. if csv.js fails to parse it, it informs the user it is an invalid
format.

2. if csv.js parses successfully but is the incorrect format for the
given selection it informs the user it is an invalid format. (e.g. file
is setup for document format, but is uploaded to passengers)

3. if csv.js parses successfully AND the file matches the format for the
selected type of import then it informs the user that it will erase all
previous entries and replace it with the uploaded files, while
presenting the user with a dialog option to accept or decline the event.