Issue #15, #17, #29, #32, #33 and improvements to rules/queries
-#15: Added md-dialog to rules/queries for a better error message
delivery for several client-side errors.

-#17: Flights only queries on >= 3 characters now

-#29: Spinner is now hidden correctly on failure paths

-#32: Fixed issue with datepicker not accepting non-date objects. An
attempt is made to convert string from the ui grid to a date object
before passing it to the datepicker.

-#33: Removed format constraints on password login box per request.

Improvements to rules/queries: 

-Canceling and clicking on the backdrop to close the side-nav results in
a cancel event that clears erroneous data so that it doesn't retain in
most situations upon reopening the same side-nav.

-Copy no longer immediately copies the selected rule. The user is now
prompted to save the copy as if it were a new rule.

-Updating an existing rule no longer allows you to change the start date

-Changed the query side-nav to "md-sidenav-right" class as it operates
much faster than "left" for some reason.