Rule Update bug, Start/End date inconsistency, watchlist import improved
-RulePersistenceService, RulePersistenceServiceImpl, and UdrServiceImpl
new overloaded method was added to circumvent an EntityExistsException.
update(rule, User) is used if and only if the author of the rule to be
updated is the same as the current editor for the update. This prevents
the fetchUser() method from being called and creating a series of
detached filter objects that collide with the existing filter nested
within the rule to be updated that gets pulled prior in session.

-FlightsController.js, PaxController.js both have two new watch
statements added in order to insure that the start date and end date on
either Flights or Passengers screen are held in appropriate format. I.e.
Start Date's clock time = 00:00:00 End Date's clock time = 23:59:59

-WatchListController.js had speed improvements on validation for
importing CSV files. Also, expanded feature for catching and displaying
invalid date formats to the user, giving the particular issue with the
imported format and the index line location of the offending format.